+++ src/Utils.php
@@ -9,28 +9,66 @@ trait Utils
 	protected function calculateTextCoordinates( $text, $size, $position, $color, $angle, $font )
 	{
 		// 
-		$bbox = $this->imageTTFBBoxExtended( $size, $angle, $font, $text );
+		$realBBox = $this->imageTTFBBoxExtended( $size, $angle, $font, $text );
 
 		//
-		$customWidth = $bbox[ 'width' ]; /*abs( $bbox[ 0 ] - $bbox[ 4 ] );*/
-		$customHeight = $bbox[ 'height' ]; /*abs( $bbox[ 1 ] - $bbox[ 5 ] );*/
+		$customWidth = $realBBox[ 'width' ]; /*abs( $realBBox[ 0 ] - $realBBox[ 4 ] );*/
+		$customHeight = $realBBox[ 'height' ]; /*abs( $realBBox[ 1 ] - $realBBox[ 5 ] );*/
 
 		//
-		$x = $bbox[ 'x' ] + ( ( is_array( $position ) && !empty( $position[ 'x' ] ) ) ? $position[ 'x' ] : 0 );
-		$y = $bbox[ 'y' ] + ( ( is_array( $position ) && !empty( $position[ 'y' ] ) ) ? $position[ 'y' ] : 0 );
+		$x = $realBBox[ 'x' ] + ( ( is_array( $position ) && !empty( $position[ 'x' ] ) ) ? $position[ 'x' ] : 0 );
+		$y = $realBBox[ 'y' ] + ( ( is_array( $position ) && !empty( $position[ 'y' ] ) ) ? $position[ 'y' ] : 0 );
+
+		//
+		$horizontalPadding = ( $this->horizontalPadding + $this->borderSize );
+		$verticalPadding = ( $this->verticalPadding + $this->borderSize );
+
+		//
+		// $x += ( $this->horizontalPadding + $this->borderSize );
+		// $y += ( $this->verticalPadding + $this->borderSize );
 
 		// Clear memory
-		$bbox = null;
+		$realBBox = null;
 
+		//
+		$fontHeight = $this->calculateFontHeight( $text, $font, $size, $angle );
+
+		//
 		return [
-			'width'		=> $customWidth,
-			'height'	=> $customHeight,
-			'x'			=> $y, // ?
-			'top'		=> $y,
-			'bottom'	=> $y + $customHeight,
-			'y'			=> $x, // ?
-			'left'		=> $x,
-			'right'		=> $x + $customWidth,
+			'width'				=> $customWidth,
+			'height'			=> $customHeight,
+			'y'					=> $y,
+			'top'				=> $y,
+			'bottom'			=> $y + $customHeight,
+			'x'					=> $x,
+			'left'				=> $x,
+			'right'				=> $x + $customWidth,
+			'lowerCharHeight' 	=> $fontHeight[ 'lower' ],
+			'upperCharHeight' 	=> $fontHeight[ 'upper' ],
+			'charHeightDiffer'	=> $fontHeight[ 'differ' ],
+		];
+	}
+
+	/**
+	 * 
+	 */
+	protected function calculateFontHeight( $text, $font, $size, $angle )
+	{
+		// 
+		$lowerCharBBox = imageTTFBBox( $size, $angle, $font, mb_strtolower( $text ) );
+		if( $lowerCharBBox[ 3 ] > 0 ) { $lowerCharHeight = abs( $lowerCharBBox[ 7 ] - $lowerCharBBox[ 1 ] ) - 1; }
+		else { $lowerCharHeight = abs( $lowerCharBBox[ 7 ] ) - abs( $lowerCharBBox[ 1 ] ); }
+
+		//
+		$upperCharBBox = imageTTFBBox( $size, $angle, $font, mb_strtoupper( $text ) );
+		if( $upperCharBBox[ 3 ] > 0 ) { $upperCharHeight = abs( $upperCharBBox[ 7 ] - $upperCharBBox[ 1 ] ) - 1; }
+		else { $upperCharHeight = abs( $upperCharBBox[ 7 ] ) - abs( $upperCharBBox[ 1 ] ); }
+
+		//
+		return [
+			'lower' 	=> $lowerCharHeight,
+			'upper' 	=> $upperCharHeight,
+			'differ'	=> $upperCharHeight - $lowerCharHeight
 		];
 	}
 
@@ -42,28 +80,33 @@ trait Utils
 		// Minimal height calculation
 		$patternBBox = imageTTFBBox( $size, $angle, $fontFile, 'AMWpq' );
 
-		/*this function extends imagettfbbox and includes within the returned array
+		/* 
+		This function extends imagettfbbox and includes within the returned array
 		the actual text width and height as well as the x and y coordinates the
 		text should be drawn from to render correctly.  This currently only works
-		for an angle of zero and corrects the issue of hanging letters e.g. jpqg*/
-		$bbox = imageTTFBBox( $size, $angle, $fontFile, $text );
+		for an angle of zero and corrects the issue of hanging letters e.g. jpqg
+		*/
+		$realBBox = imageTTFBBox( $size, $angle, $fontFile, $text );
+
+		//
+		$resultBBox = [ ];
 	
-		// calculate x baseline
-		if( $bbox[ 0 ] >= -1 ) { $bbox[ 'x' ] = abs( $bbox[ 0 ] + 1 ) * -1; } 
-		else { $bbox[ 'x' ] = abs( $bbox[ 0 ] + 2 ); }
+		// Calculate x baseline
+		if( $realBBox[ 0 ] >= -1 ) { $resultBBox[ 'x' ] = abs( $realBBox[ 0 ] + 1 ) * -1; } 
+		else { $resultBBox[ 'x' ] = abs( $realBBox[ 0 ] + 2 ); }
 	
-		// calculate actual text width
-		if( $bbox[ 0 ] < -1 ) {	$bbox[ 'width' ] = abs( $bbox[ 2 ] ) + abs( $bbox[ 0 ] ) - 1; }
-		else { $bbox[ 'width' ] = abs( $bbox[ 2 ] - $bbox[ 0 ] ); }
+		// Calculate actual text width
+		if( $realBBox[ 0 ] < -1 ) {	$resultBBox[ 'width' ] = abs( $realBBox[ 2 ] ) + abs( $realBBox[ 0 ] ) - 1; }
+		else { $resultBBox[ 'width' ] = abs( $realBBox[ 2 ] - $realBBox[ 0 ] ); }
 	
-		// calculate y baseline
-		$bbox[ 'y' ] = abs( $patternBBox[ 5 ] + 1 );
+		// Calculate y baseline
+		$resultBBox[ 'y' ] = abs( $patternBBox[ 5 ] + 1 );
 	
-		// calculate actual text height
-		if( $patternBBox[ 3 ] > 0) { $bbox[ 'height' ] = abs( $patternBBox[ 7 ] - $patternBBox[ 1 ] ) - 1; }
-		else { $bbox[ 'height' ] = abs( $patternBBox[ 7 ] ) - abs( $patternBBox[ 1 ] ); }
-	
-		return $bbox;
+		// Calculate actual text height
+		if( $patternBBox[ 3 ] > 0) { $resultBBox[ 'height' ] = abs( $patternBBox[ 7 ] - $patternBBox[ 1 ] ) - 1; }
+		else { $resultBBox[ 'height' ] = abs( $patternBBox[ 7 ] ) - abs( $patternBBox[ 1 ] ); }
+
+		return $resultBBox;
 	}
 
 	/**
@@ -94,7 +137,6 @@ trait Utils
 		// a wonky rectangle as a result of rounding error). (Not shown)
 		// imageFilledPolygon( $image, [ $x1, $y1, $x2, $y2, $x3, $y3, $x4, $y4 ], 4, $color );
 
-		// Рисование многоугольника
 		/*
 		imageFilledPolygon($image, [
 			$X, $Y - $height,
@@ -143,34 +185,28 @@ trait Utils
 		$height = imagesY( $image );
 	
 		// Convert hex-values to rgb
-		for( $i = 0; $i <= 3; $i++ )
-		{
-			$colors[ $i ] = [
-				hexdec( substr( $colors[ $i ], 1, 2 ) ),
-				hexdec( substr( $colors[ $i ], 3, 2 ) ),
-				hexdec( substr( $colors[ $i ], 5, 2 ) )
-			];
-		}
+		foreach( $colors as &$color ) { $color = sscanf( $color, '#%2x%2x%2x' ); }
 
 		// Start with top left color
 		$rgb = $colors[ 0 ]; 
 
+		//
 		for( $x = 0; $x <= $width; $x++ ) 
 		{
 			for( $y = 0 ; $y <= $height; $y++ ) 
 			{
-				// set pixel color 
+				// Set pixel color 
 				$col = imageColorAllocate( $image, $rgb[0], $rgb[1], $rgb[2] );
 				imageSetPixel( $image, $x - 1, $y - 1, $col );
 
-				// calculate new color  
+				// Calculate new color  
 				for( $i = 0; $i <= 2; $i++ ) 
 				{
 					$rgb[ $i ] =
-						$colors[ 0 ][ $i ] * ( ( $width - $x ) * ( $height - $y ) / ( $width * $height ) ) +
-						$colors[ 1 ][ $i ] * ( $x * ( $height - $y ) / ( $width * $height ) ) +
-						$colors[ 2 ][ $i ] * ( ( $width - $x ) * $y  / ( $width * $height ) ) +
-						$colors[ 3 ][ $i ] * ( $x * $y / ( $width * $height ) );
+						$colors[ 0 ][ $i ] * ( ( $width - $x ) * ( $height - $y ) / ( $width * $height ) ) 
+						+ $colors[ 1 ][ $i ] * ( $x * ( $height - $y ) / ( $width * $height ) ) 
+						+ $colors[ 2 ][ $i ] * ( ( $width - $x ) * $y  / ( $width * $height ) ) 
+						+ $colors[ 3 ][ $i ] * ( $x * $y / ( $width * $height ) );
 				}
 			}
 		}
@@ -189,7 +225,9 @@ trait Utils
 			'bottom' => 0,
 			'center' => 0 
 		];
-		
+
+		$objectIndex = -1;
+
 		foreach( $this->objectList as $value ) 
 		{
 			if( array_get( $value, 'inline', false ) )
@@ -197,12 +235,13 @@ trait Utils
 				continue;
 			}
 
-			$height = $value[ 'size' ][ 'height' ] + $this->linePadding;
+			$objectIndex++;
+			$height = array_get( $value, 'size.height' ) + $this->lineSpacing;
 
 			$sectionsHeight[ 'all' ] += $height;
-			$sectionsHeight[ 'top' ] += ( empty( $value[ 'position' ][ 'y' ] ) && $value[ 'position' ][ 'vertical-alignment' ] === 'top' ) ? $height : 0;
-			$sectionsHeight[ 'bottom' ] += ( empty( $value[ 'position' ][ 'y' ] ) && $value[ 'position' ][ 'vertical-alignment' ] === 'bottom' ) ? $height : 0;
-			$sectionsHeight[ 'center' ] += ( empty( $value[ 'position' ][ 'y' ] ) && $value[ 'position' ][ 'vertical-alignment' ] === 'center' ) ? $height : 0;
+			$sectionsHeight[ 'top' ] += ( empty( array_get( $value, 'position.y' ) ) && array_get( $value, 'position.vertical-alignment' ) === 'top' ) ? $height : 0;
+			$sectionsHeight[ 'bottom' ] += ( empty( array_get( $value, 'position.y' ) ) && array_get( $value, 'position.vertical-alignment' ) === 'bottom' ) ? $height : 0;
+			$sectionsHeight[ 'center' ] += ( empty( array_get( $value, 'position.y' ) ) && array_get( $value, 'position.vertical-alignment' ) === 'center' ) ? $height : 0;
 		}
 
 		return $sectionsHeight;
@@ -308,7 +347,7 @@ trait Utils
 							$this->horizontalPadding, 
 							$centerTop, 
 							$this->size[ 'width' ] - $this->horizontalPadding, 
-							$centerTop + $sectionsHeight[ 'center' ] - $this->linePadding, 
+							$centerTop + $sectionsHeight[ 'center' ] - $this->lineSpacing, 
 							$this->createColor( $image, '#A60000' ) );
 		}