+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 0.6.5
+* Improved calculation text block size.
+* Improved line spacing.
+* `setLinePadding` rename to `setLineSpacing`.
+* Remove param `transparent` from `setBackgroundImage`.
+* Small fixes and changes.
+
 ### 0.6.0
 * Add text `style`: `underline`, `strike`.
 * Add `addSeparator` function - for horizontal separation.
@@ -5,4 +12,4 @@
 * Add multi-line text support.
 
 ### 0.5.0
-* First version
+* First version.